+++ timeRangeLib/timeRangeValidator.go
@@ -1,4 +1,4 @@
-package scheduler
+package timeRangeLib
 
 import (
 	"errors"
